+++ lumin/nn/callbacks/data_callbacks.py
@@ -1,0 +1,60 @@
+from typing import Union, Tuple, Callable, Optional
+import numpy as np
+
+from torch import Tensor
+
+from .callback import Callback
+from ..data.batch_yielder import BatchYielder
+from ..data.fold_yielder import FoldYielder
+from ...utils.misc import to_np, to_device
+from ..models.abs_model import AbsModel
+
+
+class BinaryLabelSmooth(Callback):
+    '''Apply label smoothing to binary classes, based on arXiv:1512.00567'''
+    def __init__(self, coefs:Union[float,Tuple[float,float]]=0, model:Optional[AbsModel]=None):
+        super().__init__(model=model)
+        self.coefs = coefs if isinstance(coefs, tuple) else (coefs, coefs)
+    
+    def on_epoch_begin(self, batch_yielder:BatchYielder, **kargs) -> None:
+        '''Apply smoothing at train-time'''
+        batch_yielder.targets = batch_yielder.targets.astype(float)
+        batch_yielder.targets[batch_yielder.targets == 0] = self.coefs[0]
+        batch_yielder.targets[batch_yielder.targets == 1] = 1-self.coefs[1]
+         
+    def on_eval_begin(self, targets:Tensor, **kargs) -> None:
+        '''Apply smoothing at test-time'''
+        targets[targets == 0] = self.coefs[0]
+        targets[targets == 1] = 1-self.coefs[1]
+
+
+class SequentialReweight(Callback):
+    '''During ensemble training, sequentially reweight data in validation folds based on prediction performance of last trained model.
+       Reweighting highlights data which are easier or more difficult to predict to the next model being trained'''
+    def __init__(self, reweight_func:Callable[[Tensor,Tensor], Tensor], scale:float=1e-1, model:Optional[AbsModel]=None):
+        super().__init__(model=model)
+        self.scale,self.reweight_func = scale,reweight_func
+
+    def reweight_fold(self, fy:FoldYielder, fold_id:int) -> None:
+        fld = fy.get_fold(fold_id)
+        preds = self.model.predict_array(fld['inputs'], as_np=False)
+        coefs = to_np(self.reweight_func(preds, to_device(Tensor(fld['targets']))))
+        weight = np.sum(fld['weights'])
+        fld['weights'] += self.scale*coefs*fld['weights']
+        fld['weights'] *= weight/np.sum(fld['weights'])
+        fy.foldfile[f'fold_{fold_id}/weights'][...] = fld['weights'].squeeze()
+    
+    def on_train_end(self, fy:FoldYielder, val_id:int, **kargs) -> None: self.reweight_fold(fy, val_id)
+
+
+class SequentialReweightClasses(SequentialReweight):
+    '''SequentialReweight designed for classification, which renormalises class weights to original weight-sum after reweighting'''
+    def reweight_fold(self, fy:FoldYielder, fold_id:int) -> None:
+        fld = fy.get_fold(fold_id)
+        preds = self.model.predict_array(fld['inputs'], as_np=False)
+        coefs = to_np(self.reweight_func(preds, to_device(Tensor(fld['targets']))))
+        for c in set(fld['targets'].squeeze()):
+            weight = np.sum(fld['weights'][fld['targets'] == c])
+            fld['weights'][fld['targets'] == c] += self.scale*(coefs*fld['weights'])[fld['targets'] == c]
+            fld['weights'][fld['targets'] == c] *= weight/np.sum(fld['weights'][fld['targets'] == c])
+        fy.foldfile[f'fold_{fold_id}/weights'][...] = fld['weights'].squeeze()