+++ lumin/nn/models/blocks/body.py
@@ -1,4 +1,5 @@
 from typing import Optional, Callable
+import numpy as np
 
 import torch.nn as nn
 import torch
@@ -6,7 +7,7 @@ from torch import Tensor
 
 from ..layers.activations import lookup_act
 from ..initialisations import lookup_normal_init
-
+    
 
 class FullyConnected(nn.Module):
     '''Fully connected set of hidden layers.
@@ -19,8 +20,11 @@ class FullyConnected(nn.Module):
         self.depth,self.width,self.do,self.bn,self.act,self.res,self.dense = depth,width,do,bn,act,res,dense,
         self.lookup_init,self.lookup_act,self.freeze = lookup_init,lookup_act,freeze
 
-        self.layers = nn.ModuleList([self.get_layer(d) for d in range(depth)])
-        if dense: self.layers += [self.get_layer(depth, self.width*(2**(self.depth)), self.width)]
+        if self.res:
+            self.depth = int(np.ceil(self.depth/2))  # Each block will contain 2 layers
+            self.res_bns = nn.ModuleList([nn.BatchNorm1d(self.width) for d in range(self.depth)])
+        self.layers = nn.ModuleList([self.get_layer(d) for d in range(self.depth)])
+        #if dense: self.layers += [self.get_layer(depth, self.width*(2**(self.depth)), self.width)]
         if self.freeze: self.freeze_layers
 
     def __getitem__(self, key:int) -> nn.Module: return self.layers[key]
@@ -36,15 +40,15 @@ class FullyConnected(nn.Module):
         fan_in  = width if fan_in  is None else fan_in
         fan_out = width if fan_out is None else fan_out
         
-        layers = [nn.Linear(fan_in, fan_out)]
-        self.lookup_init(self.act, fan_in, fan_out)(layers[-1].weight)
-        if self.act != 'linear': layers.append(self.lookup_act(self.act))
-        if self.bn or self.res: layers.append(nn.BatchNorm1d(fan_out))
-        if self.do: 
-            if self.act == 'selu':
-                layers.append(nn.AlphaDropout(self.do))
-            else:
-                layers.append(nn.Dropout(self.do))
+        layers = []
+        for i in range(2 if self.res else 1):
+            layers.append(nn.Linear(fan_in, fan_out))
+            self.lookup_init(self.act, fan_in, fan_out)(layers[-1].weight)
+            if self.act != 'linear': layers.append(self.lookup_act(self.act))
+            if self.bn and i == 0:  layers.append(nn.BatchNorm1d(fan_out))  # In case of residual, BN will be added after addition
+            if self.do: 
+                if self.act == 'selu': layers.append(nn.AlphaDropout(self.do))
+                else:                  layers.append(nn.Dropout(self.do))
         return nn.Sequential(*layers)
     
     def forward(self, x:Tensor) -> Tensor:
@@ -52,7 +56,12 @@ class FullyConnected(nn.Module):
             for l in self.layers[:-1]: x = torch.cat((l(x), x), 1)
             x = self.layers[-1](x)
         else:
-            for l in self.layers: x = l(x)+x if self.res else l(x)
+            for i, l in enumerate(self.layers):
+                if self.res:
+                    x = l(x)+x
+                    x = self.res_bns[i](x)  # Renormalise after addition
+                else:
+                    x = l(x)
         return x
     
-    def get_out_size(self) -> int: return self.width
+    def get_out_size(self) -> int: return self.width if not self.dense else self.width*(2**(self.depth-1))