+++ lumin/nn/models/initialisations.py
@@ -30,3 +30,4 @@ def lookup_uniform_init(act:str, fan_in:Optional[int]=None, fan_out:Optional[int
     if act == 'linear':     return nn.init.xavier_uniform_
     if 'swish' in act:      return partial(nn.init.kaiming_uniform_, nonlinearity='relu')
     raise ValueError("Activation not implemented")
+