+++ lumin/plotting/interpretation.py
@@ -43,12 +43,12 @@ def plot_embedding(embed:OrderedDict, feat:str, savename:Optional[str]=None, set
         plt.show()
 
     
-def plot_1d_partial_dependence(model:Any, df:pd.DataFrame, feat:str, ignore_feats:List[str]=[], input_pipe:Pipeline=None, 
+def plot_1d_partial_dependence(model:Any, df:pd.DataFrame, feat:str, ignore_feats:Optional[List[str]]=None, input_pipe:Pipeline=None, 
                                sample_sz:Optional[int]=None, weights:Optional[np.ndarray]=None,  n_clusters:int=10, n_points:int=20,
                                savename:Optional[str]=None, settings:PlotSettings=PlotSettings()) -> None:
     '''Wrapper for PDPbox to plot 1D dependence of specified feature using provided NN of RF'''
     if sample_sz is not None: df = df.sample(sample_sz, weights=weights)
-    iso = pdp.pdp_isolate(model, df, [f for f in df.columns if f not in ignore_feats], feat, num_grid_points=20)
+    iso = pdp.pdp_isolate(model, df, [f for f in df.columns if ignore_feats is None or f not in ignore_feats], feat, num_grid_points=20)
     if input_pipe is not None: _deprocess_iso(iso, input_pipe, feat, df.columns)
 
     with sns.axes_style(settings.style), sns.color_palette(settings.cat_palette):
@@ -64,13 +64,13 @@ def plot_1d_partial_dependence(model:Any, df:pd.DataFrame, feat:str, ignore_feat
         plt.show()
 
 
-def plot_2d_partial_dependence(model:Any, df:pd.DataFrame, feats:Tuple[str,str], ignore_feats:List[str]=[], input_pipe:Pipeline=None,
+def plot_2d_partial_dependence(model:Any, df:pd.DataFrame, feats:Tuple[str,str], ignore_feats:Optional[List[str]]=None, input_pipe:Pipeline=None,
                                sample_sz:Optional[int]=None, weights:Optional[np.ndarray]=None, n_points:Tuple[int,int]=[20,20],
                                savename:Optional[str]=None, settings:PlotSettings=PlotSettings()) -> None:
     '''Wrapper for PDPbox to plot 2D dependence of specified pair of features using provided NN of RF'''    
     check_pdpbox()
     if sample_sz is not None: df = df.sample(sample_sz, weights=weights)
-    interact = pdp.pdp_interact(model, df, [f for f in df.columns if f not in ignore_feats], feats, num_grid_points=n_points)
+    interact = pdp.pdp_interact(model, df, [f for f in df.columns if ignore_feats is None or f not in ignore_feats], feats, num_grid_points=n_points)
     if input_pipe is not None: _deprocess_interact(interact, input_pipe, feats, df.columns)
             
     with sns.axes_style(settings.style), sns.color_palette(settings.cat_palette):