+++ lumin/plotting/plot_settings.py
@@ -24,6 +24,7 @@ class PlotSettings:
 
         self.savepath = Path('.') if 'savepath' not in kargs else kargs['savepath']
         self.format   = '.pdf'    if 'format'   not in kargs else kargs['format']
+        if '.' not in self.format: self.format = '.' + self.format
 
         self.h_small = 4  if 'h_small' not in kargs else kargs['h_small']
         self.h_mid   = 8  if 'h_mid'   not in kargs else kargs['h_mid']